requirements.txt
blinker>=1.8.2
certifi>=2024.8.30
charset-normalizer>=3.3.2
click>=8.1.7
Flask>=3.0.3
idna>=3.8
itsdangerous>=2.2.0
Jinja2>=3.1.4
MarkupSafe>=2.1.5
oauthlib>=3.2.2
pillow>=10.4.0
python-dotenv>=1.0.1
pytz>=2024.1
requests>=2.32.3
requests-oauthlib>=2.0.0
urllib3>=2.2.2
Werkzeug>=3.0.4
